backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-migrate==4.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pgvector==0.3.2
requests==2.32.3
numpy==1.26.4
open-clip-torch==2.24.0
torch==2.3.0
Pillow==10.3.0
umap-learn==0.5.6